modify broken handlers that caused hanging

+++ src/handlers/update.ts
@@ -2,30 +2,31 @@ import prisma from "../db";
 
 // Get all updates
 export const getUpdates = async (req, res) => {
-  const product = await prisma.product.findUnique({
+  const products = await prisma.product.findMany({
     where: {
-      id: req.product.id,
+      belongsToId: req.user.id,
     },
     include: {
       updates: true,
     },
   });
 
-  res.json({ data: product.updates });
+  const updates = products.reduce((allUpdates, product) => {
+    return [...allUpdates, ...product.updates];
+  }, []);
+
+  res.json({ data: updates });
 };
 
 // Get one update for one product
 export const getUpdate = async (req, res) => {
-  const product = await prisma.product.findFirst({
+  const update = await prisma.update.findUnique({
     where: {
-      id: req.product.id,
-    },
-    include: {
-      updates: true,
+      id: req.params.id,
     },
   });
 
-  res.json({ data: product.updates });
+  res.json({ data: update });
 };
 
 // Create an update
@@ -41,11 +42,11 @@ export const createUpdate = async (req, res) => {
   }
 
   const update = await prisma.update.create({
-      data: {
-          title: req.body.title,
-          body: req.body.body,
-          product: {connect: product}
-    }
+    data: {
+      title: req.body.title,
+      body: req.body.body,
+      product: { connect: product },
+    },
   });
 
   res.json({ data: update });
@@ -84,21 +85,30 @@ export const updateUpdate = async (req, res) => {
 
 // Delete an update
 export const deleteUpdate = async (req, res) => {
-  const { updateId } = req.params;
-
-  await prisma.update.findUnique({
-    where: { id: updateId },
+  const products = await prisma.product.findMany({
+    where: {
+      belongsToId: req.user.id,
+    },
     include: {
-      product: {
-        select: {
-          belongsToId: true,
-        },
-      },
+      updates: true,
     },
   });
 
+  const updates = products.reduce((allUpdates, product) => {
+    return [...allUpdates, ...product.updates];
+  }, []);
+
+  const match = updates.find((update) => update.id === req.params.id);
+
+  if (!match) {
+    // handle this
+    return res.json({ message: "nope" });
+  }
+
   const deleted = await prisma.update.delete({
-    where: { id: updateId },
+    where: {
+      id: req.params.id,
+    },
   });
 
   res.json({ data: deleted });
